requirements.txt
deepspeed==0.4.0
numpy==1.24.1
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==10.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
protobuf==3.20.*
tensorboardX==1.8
timm==0.4.12
torch==2.0.0
torchmetrics==0.7.3
torchscale==0.2.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
sentencepiece
